SCRUM-8 Add UserPolicy and User,Book,UserController tests

+++ routes/web.php
@@ -8,11 +8,11 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/users/{user}/books', [UserController::class, 'showBooks']);
+Route::get('/users/{user}/books', [UserController::class, 'showBooks'])->middleware('auth');
 
-Route::get('/users/{user}/books/onloan', [UserController::class, 'showBooksOnLoan']);
-Route::get('/users/{user}/books/onloan/create', [UserController::class, 'booksOnLoanCreate']);
-Route::post('/users/{user}/books/{book}/onloan', [UserController::class, 'addBookOnLoan']);
-Route::delete('/users/{user}/books/{book}/onloan', [UserController::class, 'removeBookOnLoan']);
+Route::get('/users/{user}/books/onloan', [UserController::class, 'showBooksOnLoan'])->middleware('auth');
+Route::get('/users/user/books/onloan/create', [UserController::class, 'booksOnLoanCreate'])->middleware('auth');
+Route::post('/users/{user}/books/{book}/onloan', [UserController::class, 'addBookOnLoan'])->middleware('auth')->can('addBookOnLoan','book');
+Route::delete('/users/{user}/books/{book}/onloan', [UserController::class, 'removeBookOnLoan'])->middleware('auth')->can('removeBookOnLoan,book');
 
 Route::get('/books/search', [BookController::class, 'search']);
